# Use CryptoStream to convert stream from base64